backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20.0,<2.0.0
psycopg2-binary>=2.8
Pillow

